Add additional refactors to move lengthy facade methods into modules
