+++ app/models/favorite.rb
@@ -1,4 +1,6 @@
 class Favorite < ApplicationRecord
+  include DataUsaClientHelper
+
   belongs_to :user
   belongs_to :location
 
@@ -6,4 +8,12 @@ class Favorite < ApplicationRecord
 
   scope :active, ->{ where(active: true) }
   scope :eliminated, ->{ where(active: false) }
+
+  def city_name
+    city_overview[:display_name]
+  end
+
+  def city_url
+    city_overview[:url_name]
+  end
 end
