+++ app/views/users/show.html.erb
@@ -7,12 +7,12 @@
     <li class="breadcrumb-item active">Dashboard</li>
   </ol>
 </nav>
-<h1>Welcome <%= @user.to_s %></h1>
+<h1>Welcome <%= current_user.name %></h1>
 <h4>Active Favorites: <%= current_user.favorites.count %></h4>
 
 <div class="list-group">
   <% current_user.favorites.all.each do |favorite| %>
-    <%= link_to "/city?q=#{favorite.slug}&id=#{favorite.geo_id}",
+    <%= link_to "/locations?q=#{favorite.city_name}&id=#{favorite.url_name}",
                 class: 'list-group-item list-group-item-action favorites' do %>
       <div class="d-flex w-100 justify-content-between">          
         <h5 class="mb-1"><%= favorite.name %></h5>
